requirements.txt
antlr4-python3-runtime==4.9.3
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
hydra-core==1.3.2
idna==3.11
omegaconf==2.3.0
packaging==25.0
PyYAML==6.0.3
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
